requirements.txt
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Web Framework
streamlit==1.29.0

# Data Format
pyyaml==6.0.1

# Optional: for better performance
joblib==1.3.2

# Additional utilities
python-dateutil==2.8.2
